refactor 💥: Update and organise event bases (#636)
